Add 6 Files for Manipulating Tools
This commit adds 6 files for operating on tools in the MySQL
database. Here is a list of all 6 new files and a short description.

CreateEmptyTool.php:
Creates a new tool with default values.

CreateToolLink.php:
Creates a new reference link for a given tool_uuid. A tool
can have as many associated links as you'd like.

DeleteTool.php:
Deletes a tool from the database. As I'm typing this, I'm realizing
I should probably modify this to remove all links associated with
the removed tool from tbl_links.

DeleteToolLink.php:
Deletes a particular link associated with a tool.

UpdateTool.php:
Updates a single column for a particular tool.

UpdateToolLink:
Updates a single column for a particular tool in tbl_links.