Operate on Links Based on link_id
The tbl_links table in the database now has a primary key column
for each link. This will greatly simplify the React code on the
frontend.

GetToolById and GetAllTools now return the link_ids of all links
with the rest of the tool data.

UpdateToolLink.php and DeleteToolLink.php now use the link_id
to operate on links instead of combining tool_uuid and link
columns.