Add GetNextLinkId.php for Getting the Next Primary Key for a Key
This commit adds GetNextLinkId.php which runs the following
query to get the next link's primary key. Because, remember
the link_id (primary key of tbl_links) is an auto incremented
number.

SELECT AUTO_INCREMENT as link_id
FROM information_schema.TABLES
WHERE TABLE_SCHEMA = 'toolbox' AND TABLE_NAME = 'tbl_links';

This endpoint is required by the "Links" section of the
React app, because the application has to know what the
next link_id is when it renders an array of links. In other
words it used link_id as a key for rendering links.