Allow Empty Strings on Update AND Change INSERT INTO
This commit allows empty strings to be stored in both
tbl_tool and tbl_links tables. Before, there was a check
if the value getting set was not null which was preventing
empty strings from getting stored properly.

Also, UpdateTool.php and UpdateToolLink.php INSERT INTO
statements are modified to make use of the default values
of both tables.